+++ src/LineChart/actions.js
