+++ src/LineChart/component.js
