+++ src/LineChart/reducer.js
