+++ src/LineChart/selectors.js
